requirements.txt
pysondb
PyYAML
requests